feat: sidebar lists work in progress
